Rev.board 2.0
Add: NTC read, input volt read, button_analog read, Tasks add(1ms, 10ms, 15ms, 190ms, 500ms)

+++ Read_Battery.h
@@ -2,7 +2,7 @@
 #define _Read_Battery_H_
 
 void Read_Input_Volt();
-void Check_Input_Voltage(double check_volt);
-double GetInputVolt();
+void Check_Input_Voltage(int check_volt);
+int GetInputVolt();
 
 #endif
